ffi: Move Ruma wrappers / extension traits to new module
