+++ bindings/matrix-sdk-ffi/src/lib.rs
@@ -34,6 +34,7 @@ mod room;
 mod room_info;
 mod room_list;
 mod room_member;
+mod ruma;
 mod session_verification;
 mod sync_service;
 mod task_handle;
@@ -49,8 +50,8 @@ use matrix_sdk_ui::timeline::{BackPaginationStatus, EventItemOrigin};
 
 use self::{
     error::ClientError,
+    ruma::{MediaSourceExt, Mentions, RoomMessageEventContentWithoutRelationExt},
     task_handle::TaskHandle,
-    timeline::{MediaSourceExt, Mentions, RoomMessageEventContentWithoutRelationExt},
 };
 
 uniffi::include_scaffolding!("api");
